refactor[#10]: 회원 정보 수정 및 탈퇴 EditProfile 구조 분리
